Backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
Flask-Limiter==3.7.0
marshmallow==3.21.3
psycopg[binary]==3.2.1
python-dotenv==1.0.1
cryptography==43.0.1
